Fix bug with the duplicated empty list when open new page
